requirements.txt
absl-py==0.8.1
alabaster==0.7.12
argh==0.26.2
asn1crypto==1.3.0
astor==0.8.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
bert-for-tf2==0.13.4
bleach==3.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.3.0
colorama==0.4.3
cryptography==2.8
cvxopt==1.2.3
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
docutils==0.16
entrypoints==0.3
flake8==3.7.9
future==0.18.2
gast==0.2.2
google-pasta==0.1.8
grpcio==1.16.1
h5py==2.10.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.14.1
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.0
kiwisolver==1.2.0
langdetect==1.0.7
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.4
mkl==2019.0
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.4
neurokit==0.2.8
nltk==3.4.5
numba==0.39.0
numpy==1.18.1
numpydoc==0.9.2
opt-einsum==3.1.0
packaging==20.1
pandas==1.0.1
pandocfilters==1.4.2
paramiko==2.6.0
params-flow==0.7.4
parso==0.6.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.3
protobuf==3.11.3
psutil==5.6.7
py-params==0.8.2
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
Pympler==0.7
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyreadline==2.1
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==5.3
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
requests==2.22.0
rope==0.16.0
Rtree==0.9.3
scikit-learn==0.22.1
scipy==1.4.1
sentencepiece==0.1.85
six==1.14.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==4.0.0
spyder-kernels==1.8.1
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
testpath==0.4.4
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
twitter==1.18.0
ujson==1.35
urllib3==1.25.8
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.1
wget==3.2
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
yapf==0.28.0
zipp==2.2.0
